Add DailyChallenge class to get & store challenges
- Ecapsulates logic for fetching daily challenges from a server,
saving, loading and calculating its expiration.

- Expiration calculation is local time agnostic. This is achieved
by noting local and server fetch times. The difference (in secs) is
used to convert local time back into server time and then used to check
against expiration time. In otherwords, the server is the source of
truth for determining if a challenge is expired.

- The expiration logic also contains a basic time manipulation check.
If device time is changed to a time before the last stored local fetch
time, the challenge is considered expired. This is done so a new fetch
can be made to resynchronize local and remote fetch times.